Feat/linkding (#4)
* fix(kustomization): correct apiVersion to v1beta1

The apiVersion in the kustomization.yaml file was incorrect. It has been
updated from v1beta to v1beta1 to ensure compatibility with the kustomize
configuration.

* chore(release): 0.0.1

* fix(linkding): update apiVersion in apps/base/linkding/kustomization.yaml

Updated the apiVersion in the kustomization.yaml file for the linkding app
from v1beta to v1beta1. This change ensures compatibility with the latest
kustomize configuration requirements.

* chore(release): 0.0.2